Update to match latest changes in JavaCPP

+++ src/main/java/org/bytedeco/javacv/ProCamTransformer.java
@@ -315,7 +315,7 @@ public class ProCamTransformer implements ImageTransformer {
             kernelData.dstDstDot(data[i].dstDstDot);
         }
 
-        int fullCapacity = kernelData.capacity();
+        long fullCapacity = kernelData.capacity();
         kernelData.capacity(data.length);
         multiWarpColorTransform(kernelData, roi, getFillColor());
         kernelData.capacity(fullCapacity);
