+++ src/main/java/org/bytedeco/javacv/ProjectiveColorTransformer.java
@@ -161,7 +161,7 @@ public class ProjectiveColorTransformer extends ProjectiveTransformer {
             kernelData.dstDstDot(data[i].dstDstDot);
         }
 
-        int fullCapacity = kernelData.capacity();
+        long fullCapacity = kernelData.capacity();
         kernelData.capacity(data.length);
         multiWarpColorTransform(kernelData, roi, getFillColor());
         kernelData.capacity(fullCapacity);
